Reworked a view more things
